+++ src/Filesystem/Generic.php
@@ -7,7 +7,10 @@ use Innmind\OperatingSystem\{
     Filesystem,
     Config,
 };
-use Innmind\Filesystem\Adapter;
+use Innmind\Filesystem\{
+    Adapter,
+    File\Content,
+};
 use Innmind\Url\Path;
 use Innmind\Server\Control\Server\Processes;
 use Innmind\FileWatch\{
@@ -15,7 +18,13 @@ use Innmind\FileWatch\{
     Factory,
     Watch,
 };
-use Innmind\Immutable\Maybe;
+use Innmind\Stream\Bidirectional;
+use Innmind\Immutable\{
+    Maybe,
+    Sequence,
+    Str,
+    Predicate\Instance,
+};
 
 final class Generic implements Filesystem
 {
@@ -32,6 +41,9 @@ final class Generic implements Filesystem
         $this->mounted = new \WeakMap;
     }
 
+    /**
+     * @internal
+     */
     public static function of(Processes $processes, Config $config): self
     {
         return new self($processes, $config);
@@ -95,5 +107,28 @@ final class Generic implements Filesystem
     public function watch(Path $path): Ping
     {
         return ($this->watch)($path);
+    }
+
+    public function temporary(Sequence $chunks): Maybe
+    {
+        $temporary = $this
+            ->config
+            ->streamCapabilities()
+            ->temporary()
+            ->new();
+
+        $temporary = $chunks->reduce(
+            Maybe::just($temporary),
+            static fn(Maybe $temporary, $chunk) => Maybe::all($temporary, $chunk)->flatMap(
+                static fn(Bidirectional $temporary, Str $chunk) => $temporary
+                    ->write($chunk->toEncoding(Str\Encoding::ascii))
+                    ->maybe()
+                    ->keep(Instance::of(Bidirectional::class)),
+            ),
+        );
+
+        return $temporary
+            ->map($this->config->io()->readable()->wrap(...))
+            ->map(Content::io(...));
     }
 }